build(deps): bump the vestad-deps group in /vestad with 6 updates (#387)
* build(deps): bump the vestad-deps group in /vestad with 6 updates

Bumps the vestad-deps group in /vestad with 6 updates:

| Package | From | To |
| --- | --- | --- |
| [axum-server](https://github.com/programatik29/axum-server) | `0.7.3` | `0.8.0` |
| [tokio-tungstenite](https://github.com/snapview/tokio-tungstenite) | `0.26.2` | `0.29.0` |
| [rcgen](https://github.com/rustls/rcgen) | `0.13.2` | `0.14.7` |
| [rand](https://github.com/rust-random/rand) | `0.9.4` | `0.10.1` |
| [nix](https://github.com/nix-rust/nix) | `0.29.0` | `0.31.2` |
| [bollard](https://github.com/fussybeaver/bollard) | `0.18.1` | `0.20.2` |


Updates `axum-server` from 0.7.3 to 0.8.0
- [Release notes](https://github.com/programatik29/axum-server/releases)
- [Changelog](https://github.com/programatik29/axum-server/blob/master/CHANGELOG.md)
- [Commits](https://github.com/programatik29/axum-server/compare/v0.7.3...v0.8.0)

Updates `tokio-tungstenite` from 0.26.2 to 0.29.0
- [Changelog](https://github.com/snapview/tokio-tungstenite/blob/master/CHANGELOG.md)
- [Commits](https://github.com/snapview/tokio-tungstenite/compare/v0.26.2...v0.29.0)

Updates `rcgen` from 0.13.2 to 0.14.7
- [Release notes](https://github.com/rustls/rcgen/releases)
- [Commits](https://github.com/rustls/rcgen/compare/v0.13.2...v0.14.7)

Updates `rand` from 0.9.4 to 0.10.1
- [Release notes](https://github.com/rust-random/rand/releases)
- [Changelog](https://github.com/rust-random/rand/blob/master/CHANGELOG.md)
- [Commits](https://github.com/rust-random/rand/compare/0.9.4...0.10.1)

Updates `nix` from 0.29.0 to 0.31.2
- [Changelog](https://github.com/nix-rust/nix/blob/master/CHANGELOG.md)
- [Commits](https://github.com/nix-rust/nix/compare/v0.29.0...v0.31.2)

Updates `bollard` from 0.18.1 to 0.20.2
- [Release notes](https://github.com/fussybeaver/bollard/releases)
- [Changelog](https://github.com/fussybeaver/bollard/blob/master/RELEASE.md)
- [Commits](https://github.com/fussybeaver/bollard/compare/v0.18.1...v0.20.2)

---
updated-dependencies:
- dependency-name: axum-server
  dependency-version: 0.8.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: vestad-deps
- dependency-name: tokio-tungstenite
  dependency-version: 0.29.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: vestad-deps
- dependency-name: rcgen
  dependency-version: 0.14.7
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: vestad-deps
- dependency-name: rand
  dependency-version: 0.10.1
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: vestad-deps
- dependency-name: nix
  dependency-version: 0.31.2
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: vestad-deps
- dependency-name: bollard
  dependency-version: 0.20.2
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: vestad-deps
...


* fix(vestad): adapt to bollard 0.20 API

- Options types moved to bollard::query_parameters
- bollard::container::Config renamed to bollard::models::ContainerCreateBody
- CreateContainerOptions.name is now Option<String> (no longer generic)
- StartContainerOptions no longer generic
- StopContainerOptions.t, RestartContainerOptions.t are Option<i32> with a new signal field
- RemoveContainerOptions has v/force/link fields
- ListContainersOptions/ListImagesOptions filters now Option<HashMap<String,Vec<String>>>
- TagImageOptions.repo/tag are Option<String>
- build_image/upload_to_container take BodyType via bollard::body_full/body_stream helpers
- import_image_stream expects Stream<Item = Result<Bytes, E>>
- BuildInfo.error replaced by error_detail.message


---------

+++ vestad/src/backup.rs
@@ -17,7 +17,7 @@ pub const DEFAULT_RETENTION_WEEKLY: usize = 2;
 pub const DEFAULT_RETENTION_MONTHLY: usize = 1;
 const MIN_DISK_SPACE_BYTES: u64 = 1_000_000_000; // 1 GB
 const DISK_SPACE_MARGIN_BYTES: u64 = 500_000_000; // 500 MB margin above container size
-pub const BACKUP_STOP_TIMEOUT_SECS: i64 = 30;
+pub const BACKUP_STOP_TIMEOUT_SECS: i32 = 30;
 pub const MIN_AGE_FOR_BACKUP_SECS: u64 = 6 * 3600;
 
 /// Acquire an exclusive file lock for the given agent. The lock is held for the
